Refactor main, app and  manager classes
