+++ books_manager.rb
@@ -9,12 +9,7 @@ class BooksManager
     @books.map { |book| puts "Title: '#{book.title}', Author: '#{book.author}'" }
   end
 
-  def create_book
-    puts "Please enter the book's"
-    print 'Title: '
-    title = gets.chomp
-    print 'Author: '
-    author = gets.chomp
+  def create_book(title, author)
     new_book = Book.new(title, author)
     @books.push(new_book)
     print "The book '#{new_book.title}' by '#{new_book.author}' was added successfully"
